fix: refrain from global console mod
KISS and just do a colored .log message
